Get rid of the 'hi' code
These code segments is of questionable use. Also, using absolute values such as pixels can be problematic in cases where a theme is used that does not have bounding boxes, and a person for example posts a discord help url. It will generate a very big scroll bar that will confuse and frustrate many people. In situations like that, it is more ideal to use calculations `calc()` alongside percentages of the width of the display window (vw).